Game Server: Created tests for EloAlgorithm and they work perfectly

+++ src/main/java/gameserver/control/RatingAlgorithm.java
@@ -7,13 +7,13 @@ public interface RatingAlgorithm {
      * @param loserRating Rating of the loser of a match before a match is started
      * @return The change to the losers's rating
      */
-    int calculateLoserChange( double winnerRating, double loserRating );
+    int calculateLoserChange( int winnerRating, int loserRating );
 
     /**
      * @param winnerRating Rating of the winner of a match, before a match is started
      * @param loserRating Rating of the loser of a match before a match is started
      * @return The change to the winner's rating
      */
-    int calculateWinnerChange( double winnerRating, double loserRating );
+    int calculateWinnerChange( int winnerRating, int loserRating );
 
 }
